+++ packages/pluggableWidgets/intro-screen-native/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   We migrated from using the native FlatList to @shopify/flash-list.
+
 ## [4.2.1] - 2026-1-19
 
 ### Fixed